[EI-227] [web] refactor: Refactor file path
