+++ web/app/ui/pages/workspace/side-bar/custom-forecast-indicator-list-container.tsx
@@ -1,6 +1,6 @@
 import { SWRConfig } from 'swr';
-import ClientDataSuspense from '../../components/util/client-data-suspense';
-import CustomForecastIndicatorList from '../../components/numerical-guidance/custom-forecast-indicator/custom-forecast-indicator-list/custom-forecast-indicator-list';
+import ClientDataSuspense from '../../../components/util/client-data-suspense';
+import CustomForecastIndicatorList from '../../../components/numerical-guidance/custom-forecast-indicator/custom-forecast-indicator-list/custom-forecast-indicator-list';
 
 export default function CustomForecastIndicatorListContainer() {
   return (
