+++ web/app/ui/pages/workspace/side-bar/indicator-list-container.tsx
@@ -1,6 +1,6 @@
-import IndicatorList from '../../components/numerical-guidance/indicator/indicator-list';
-import Icon from '../../components/view/atom/icons/variant-icon';
-import Accordion from '../../components/view/molecule/accordion';
+import IndicatorList from '../../../components/numerical-guidance/indicator/indicator-list';
+import Icon from '../../../components/view/atom/icons/variant-icon';
+import Accordion from '../../../components/view/molecule/accordion';
 import { ChartSquareBarIcon } from '@heroicons/react/solid';
 
 export default function IndicatorListContainer() {
