+++ web/app/ui/pages/workspace/side-bar/metadata-list-container.tsx
@@ -1,6 +1,6 @@
 import { SWRConfig } from 'swr';
-import MetadataList from '../../components/numerical-guidance/indicator-board-metadata/metadata-list/metadata-list';
-import ClientDataSuspense from '../../components/util/client-data-suspense';
+import MetadataList from '../../../components/numerical-guidance/indicator-board-metadata/metadata-list/metadata-list';
+import ClientDataSuspense from '../../../components/util/client-data-suspense';
 
 export default function MetadataListContainer() {
   return (
