+++ web/app/ui/pages/workspace/side-bar/sidenav.tsx
@@ -3,7 +3,7 @@ import { useState } from 'react';
 import IndicatorBoardToolbar from './indicator-board-toolbar';
 import { Sidebar } from 'react-pro-sidebar';
 import { ChevronDoubleLeftIcon } from '@heroicons/react/solid';
-import IconButton from '../../components/view/atom/icons/icon-button';
+import IconButton from '../../../components/view/atom/icons/icon-button';
 
 export default function SideNav() {
   const [collapsed, setCollapsed] = useState(false);
