+++ web/app/workspace/layout.tsx
@@ -1,4 +1,4 @@
-import SideNav from '../ui/pages/workspace/sidenav';
+import SideNav from '../ui/pages/workspace/side-bar/sidenav';
 import FloatingComponentContainer from '../ui/pages/workspace/floating-component-container';
 
 export default function Layout({ children }: { children: React.ReactNode }) {
